Remove unnessesary bgfx stuff

+++ Engine/src/Engine/Core/Window.cpp
@@ -102,7 +102,7 @@ void __XXECS::Window::SetFullscreen(const bool fullscreen)
 	}
 	else if (m_isFullscreen)
 	{
-		glfwSetWindowMonitor(m_window, nullptr, m_oldPosX, m_oldPosY, m_oldWidth, m_oldHeight, 0);
+		glfwSetWindowMonitor(m_window, nullptr, static_cast<int>(m_oldPosX), static_cast<int>(m_oldPosY), static_cast<int>(m_oldWidth), static_cast<int>(m_oldHeight), 0);
 		m_isFullscreen = false;
 	}
 }
@@ -135,7 +135,7 @@ std::pair<float, float> __XXECS::Window::GetSize()
 	m_width = width;
 	m_height = height;
 
-	return {width, height};
+	return {static_cast<float>(width), static_cast<float>(height)};
 }
 
 float __XXECS::Window::GetWidth()
@@ -156,7 +156,7 @@ std::pair<float, float> __XXECS::Window::GetPos() const
 
 	glfwGetWindowPos(m_window, &x, &y);
 
-	return {x, y};
+	return { static_cast<float>(x), static_cast<float>(y)};
 }
 
 float __XXECS::Window::GetPosX() const
